Feature/dashboard and refactoring

+++ src/main/java/com/groovify/vinylshopapi/dtos/EmployeeResponseDTO.java
@@ -5,13 +5,15 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public class EmployeeResponseDTO extends UserResponseDTO {
     private String jobTitle;
-    private Double salary;
+    private BigDecimal salary;
     private Integer workHours;
     private AddressResponseDTO address;
 }
